refactor(comparison): extract row querying
